torch_ort_inference/requirements.txt
# runtime requirements.
# This list contains default required packages that will be installed via pip.
# In an environment where a requirement is already met, the package will not be updated.
# If a package of a specific version is needed, the package can be installed
# with version being specified (see Readme.md).
onnxruntime-openvino>=1.11.0
torch>=1.8
